pyproject.toml
[tool.poetry]
name = "djangoproject"
version = "0.1.0"
description = "Django testbed"
authors = ["Antoine van der Horst"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2"
pdfplumber = "^0.5.28"
tika = "^1.24"
pikepdf = "^3.2.0"
"pdfminer.six" = "20200517"
pdftotext = "^2.2.1"
PyPDF2 = "^1.26.0"
PyMuPDF = "^1.19.1"
seaborn = "^0.11.2"

[tool.poetry.scripts]
benchmark = "pdf_bench.pdf_benchmark:generate_benchmark"

[tool.poetry.dev-dependencies]
black = "^21.9b0"
pytest = "^6.2.5"
pytest-django = "^4.4.0"
rich = "^10.12.0"
django-extensions = "^3.1.3"
pre-commit = "^2.15.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "djangoProject.settings"

[tool.black]
line-length=120
